+++ CHANGELOG.md
@@ -5,9 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-[Unreleased]: https://github.com/zombiezen/go-sqlite/compare/v1.1.2...main
+[Unreleased]: https://github.com/zombiezen/go-sqlite/compare/v1.2.0...main
 
-## [Unreleased][]
+## [1.2.0][] - 2024-03-27
+
+Version 1.2.0 adds a `sqlitex.Pool.Take` method
+and improves error messages.
+
+[1.2.0]: https://github.com/zombiezen/go-sqlite/releases/tag/v1.2.0
 
 ### Added